Preparing for automated builds.
There is a top level Makefile now that builds both "kubedrutil" and
"kubedr". It can be used as follows:

$ make build BUILD_VERSION=0.42

Images are built with internal registry tag for now. To push images:

$ make pushimage

Also, I removed Prometheus configuration as we can't assume that
Prometheus would have been installed in the target cluster.