Added first automated test.
It adds a backup location and verifies that repo is successfully
initialized. The test assumes that a cluster is available and kubedr is
installed. All the env details must be provided in the file
config/testenv.json. Check sample file there for details. More details
will be documented soon.

I also fixed couple of lint errors as well as one null reference.