Skip assertion of restored files when running as part of CI.
Restore test creates a HostPath type of PV and after restore is
complete, it verifies that etcd snapshot and certificate files are
indeed restored. This check is not working when the test runs in CI
environment because the Kubernetes cluster is remotely accessed.

Until we figure out how to check remote directories, the file assertion
is skipped for CI tests. Note that the test still verifies that restore
operation is marked as success.